requirements.txt
aiofiles==24.1.0
aioshutil==1.5
annotated-types==0.7.0
anyio==4.4.0
apify==1.7.2
apify_client==1.6.4
apify_shared==1.1.2
async-timeout==4.0.3
azure-core==1.30.2
azure-cosmos==4.7.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.112.0
FLAML==2.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
numpy==1.26.4
openai==1.40.6
packaging==24.1
psutil==6.0.0
pyautogen==0.2.34
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
starlette==0.37.2
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
